[fix/navigation] refactor
refactor navigation to scale and separate contexts
