+++ buildSystem/gradle/common-app-base.gradle
@@ -100,7 +100,6 @@ android {
 dependencies {
     implementation project(Modules.SHARED_CORE)
     implementation project(Modules.SHARED_KOTLIN)
-    implementation project(Modules.SHARED_ROUTER)
 
     implementation libs.androidx.appcompat
     implementation libs.koin.android
