+++ buildSystem/gradle/common-app-signing.gradle
@@ -33,7 +33,6 @@ dependencies {
     // Shared Modules
     implementation project(Modules.SHARED_CORE)
     implementation project(Modules.SHARED_KOTLIN)
-    implementation project(Modules.SHARED_ROUTER)
 
     // Others
     implementation libs.androidx.appcompat
