+++ feature/authentication/src/main/java/com/alxnophis/jetpack/authentication/ui/view/AuthorizedFeature.kt
@@ -1,0 +1,14 @@
+package com.alxnophis.jetpack.authentication.ui.view
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun AuthorizedFeature(
+    userEmail: String,
+    onBack: () -> Unit,
+) {
+    AuthorizedScreen(
+        userEmail = userEmail,
+        navigateBack = onBack,
+    )
+}
