+++ feature/home/build.gradle
@@ -4,3 +4,7 @@ android {
     resourcePrefix 'home'
     namespace 'com.alxnophis.jetpack.home'
 }
+
+dependencies {
+    implementation project(Modules.SHARED_ROUTER)
+}
