+++ feature/notifications/src/main/java/com/alxnophis/jetpack/notifications/ui/navigation/NotificationsNavGraph.kt
@@ -1,17 +1,15 @@
 package com.alxnophis.jetpack.notifications.ui.navigation
 
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
+import androidx.compose.runtime.Composable
 import com.alxnophis.jetpack.notifications.di.injectNotifications
 import com.alxnophis.jetpack.notifications.ui.view.NotificationsScreen
-import com.alxnophis.jetpack.router.screen.Route
 
-fun NavGraphBuilder.notificationsNavGraph(navController: NavController) {
-    composable<Route.Notifications> {
-        injectNotifications()
-        NotificationsScreen(
-            navigateBack = { navController.popBackStack() },
-        )
-    }
+@Composable
+fun NotificationsFeature(
+    onBack: () -> Unit,
+) {
+    injectNotifications()
+    NotificationsScreen(
+        navigateBack = onBack,
+    )
 }
