+++ feature/posts/src/main/java/com/alxnophis/jetpack/posts/ui/view/PostDetailFeature.kt
@@ -1,0 +1,16 @@
+package com.alxnophis.jetpack.posts.ui.view
+
+import androidx.compose.runtime.Composable
+import com.alxnophis.jetpack.posts.data.model.Post
+
+@Composable
+fun PostDetailFeature(
+    onBack: () -> Unit,
+    post: Post,
+) {
+    PostDetailScreen(
+        onNavigateBack = onBack,
+        post = post,
+    )
+
+}
