+++ feature/root/build.gradle
@@ -6,6 +6,7 @@ android {
 }
 
 dependencies {
+    implementation project(Modules.SHARED_ROUTER)
     implementation project(Modules.FEATURE_HOME)
     implementation project(Modules.FEATURE_AUTHENTICATION)
     implementation project(Modules.FEATURE_GAME_BALL_CLICKER)
